backend/requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
# pandas==2.1.3
# numpy==1.26.2
# scikit-learn==1.3.2
# xgboost==2.0.2
# langchain==0.1.0
# langchain-groq==0.0.1
# python-dotenv==1.0.0
# pydantic==2.5.0
# joblib==1.3.2
# websockets==12.0

fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
numpy
scikit-learn
xgboost
langchain
langchain-groq
python-dotenv
pydantic
joblib
websockets
